refactor: wrap the test files in file
